Filesys: Support big files

+++ code/filesys/filesys.cc
@@ -96,8 +96,8 @@ FileSystem::FileSystem(bool format)
 		// Second, allocate space for the data blocks containing the contents
 		// of the directory and bitmap files.  There better be enough space!
 
-		ASSERT(mapHdr->Allocate(freeMap, FreeMapFileSize));
-		ASSERT(dirHdr->Allocate(freeMap, DirectoryFileSize, true));
+		ASSERT(mapHdr->Allocate(freeMap, (unsigned int) FreeMapFileSize));
+		ASSERT(dirHdr->Allocate(freeMap, (unsigned int) DirectoryFileSize, true));
 
 		// Flush the bitmap and directory FileHeaders back to disk
 		// We need to do this before we can "Open" the file, since open
@@ -173,7 +173,7 @@ FileSystem::FileSystem(bool format)
 //	"initialSize" -- size of file to be created
 //----------------------------------------------------------------------
 
-bool FileSystem::Create(const char *name, int initialSize, bool is_directory)
+bool FileSystem::Create(const char *name, unsigned int initialSize, bool is_directory)
 {
 	if (is_directory) {
 		if (initialSize != DirectoryFileSize)
@@ -210,7 +210,7 @@ bool FileSystem::Create(const char *name, int initialSize, bool is_directory)
 				success = FALSE; // no space on disk for data
 			else {
 				success = TRUE;
-				// everthing worked, flush all changes back to disk
+				// everything worked, flush all changes back to disk
 				hdr->WriteBack(sector);
 				directory->WriteBack(directoryFile);
 				freeMap->WriteBack(freeMapFile);
